Allow initializing DynamoDbDriver with DynamoDB resource
Just as RedisProgressManager can be initialized with a
RedisConnection, allow DynamoDbDriver to be initialized with a
DynamoDbResource, instead of relying on the DynamoDB resource created
by the boto3 default session (or requiring the caller to manipulate it
beforehand).

Use the corresponding DynamoDB client to test if the necessary
DynanmoDB tables exist.